rate use float

+++ Models/Options/MilitiaOptions.cs
@@ -5,6 +5,6 @@
         public const string Militia = "Militia";
 
         public ushort IdealMilitiaPerDate { get; set; }
-        public ushort OnDutyRate { get; set; }
+        public float OnDutyRate { get; set; }
     }
 }
